backend/requirements.txt
Django==2.2.19
pillow==9.0.1
djangorestframework==3.13.1
drf-base64==2.0
django-filter==21.1
djoser==2.1.0